+++ PhotoBooth/Components/Capture Session/CameraZoom.swift
@@ -1,0 +1,6 @@
+import CoreGraphics
+
+struct CameraZoom {
+    static let min: CGFloat = 1.0
+    static let max: CGFloat = 3.0
+}
