+++ PhotoBooth/Components/Capture Session/FlashStatus.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum FlashStatus {
+    case on
+    case off
+}
